requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
openai==1.51.2
langchain==0.2.16
langchain-openai==0.1.22
faiss-cpu==1.8.0
pypdf==4.3.1
sentence-transformers==3.2.0
nltk==3.9.1
streamlit==1.38.0
httpx==0.27.2
orjson==3.10.7